Go from int to float

+++ Core/Boundaries/GatheringGate.cs
@@ -20,14 +20,14 @@ namespace Core.Boundaries
 		DateTimeOffset StartTime, double Latitude, double Longitude, string FriendlyLocation,
 		DateTimeOffset? TimeEnded, GatheringState State, int GroupMinimum, int GroupMaximum, CharacterShard Character,
 		double Radius, bool IsDynamic, bool IsPendingDeletion, int NumberOfGuests,
-		int DegreeOfPrivacy, GatheringVisibility Visibility, DateTimeOffset TimeOfCreation, int Decay)
+		int DegreeOfPrivacy, GatheringVisibility Visibility, DateTimeOffset TimeOfCreation, float Decay)
 		: CoreOnlyData();
 
 	public record GatheringShard(long Id, UserShard Host, string Title, string Description,
         DateTimeOffset StartTime, double Latitude, double Longitude, string FriendlyLocation,
 		DateTimeOffset? TimeEnded, GatheringState State, int GroupMinimum, int GroupMaximum,
         double Radius, int DegreeOfPrivacy, int NumberOfGuests, float RelativeAngle,
-		GatheringVisibility Visibility, int Decay);
+		GatheringVisibility Visibility, float Decay);
 
 	public record GuestListBondPair(UserShard User, GatheringBond Bond);
